LayoutTests/svg/dom/systemLanguage-matching-expected.txt

systemLanguage matches against the same primary user language that navigator.language exposes. Parent language tags are synthesized from the primary language per the SVG2 5.6.5 implementation note.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- User language en-US: parent-tag synthesis matches en, en-US, en-Latn-US ---
PASS isRendered(createRect("en")) is true
PASS isRendered(createRect("en-US")) is true
PASS isRendered(createRect("en-Latn-US")) is true
PASS isRendered(createRect("eng")) is false
PASS isRendered(createRect("fr")) is false

--- User language zh-Hans-CN: parent-tag synthesis matches zh, zh-Hans, zh-Hans-CN ---
PASS isRendered(createRect("zh")) is true
PASS isRendered(createRect("zh-Hans")) is true
PASS isRendered(createRect("zh-Hans-CN")) is true
PASS isRendered(createRect("zh-Hant")) is true

--- Privacy: only the primary user language is matched, not secondary preferences ---
PASS isRendered(createRect("fr")) is true
PASS isRendered(createRect("fr-FR")) is true
PASS isRendered(createRect("en")) is false
PASS isRendered(createRect("en-US")) is false
PASS isRendered(createRect("ja")) is false
PASS isRendered(createRect("ja-JP")) is false
PASS successfullyParsed is true

TEST COMPLETE
